Support c00l new iOS 17.1 functionality

+++ install.sh
@@ -3,11 +3,14 @@ if [ ! -d "$THEOS/templates/ios/opa334" ]; then
 fi
 
 if [ -f "$THEOS/templates/ios/opa334/control_center_module.nic.tar" ]; then
-    echo "Removing legacy template..."
-    rm $THEOS/templates/ios/opa334/control_center_module.nic.tar
+  echo "Removing legacy template..."
+  rm $THEOS/templates/ios/opa334/control_center_module.nic.tar
 fi
 
-cp ./iphone_control_center_module-11up.nic.tar $THEOS/templates/ios/opa334
+if [ ! -f "$THEOS/vendor/templates/iphone_control_center_module-11up.nic.tar" ]; then
+  cp ./iphone_control_center_module-11up.nic.tar $THEOS/templates/ios/opa334
+fi
+
 cp ./iphone_ccsupport_module_provider-11up.nic.tar $THEOS/templates/ios/opa334
 
 cp -r ./Headers/ControlCenterUI $THEOS/include
